stmhal/sdcard: Clean before and Invalidate after DMA transfer.
The data cache must be invalidated AFTER the DMA finishes since speculative
access from the CM7 can fetch data before the DMA has finished causing
invalid data.

The clean (flush) before the DMA is also necessary since the cache eviction
process is unpredictable and could overwrite data from the DMA in physical
memory with stale data from the cache.

AN4839:
If all the lines are allocated, the cache controller runs the line eviction
process, where a line is selected (depending on replacement algorithm)
cleaned/invalidated, and reallocated. The data cache and Instruction cache
implement a pseudo-random replacement algorithm.

If the software is using cacheable memory regions for the DMA source/or
destination buffers. The software must trigger a cache clean before
starting a DMA operation to ensure that all the data are committed to the
subsystem memory. After the DMA transfer complete, when reading the data
from the peripheral, the software must perform a cache invalidate before
reading the DMA updated memory region.

Signed-off-by: Peter Züger <zueger.peter@icloud.com>